requirements.txt
flask
flask-cors
python-dotenv
openai
elevenlabs
langgraph-sdk
langchain-core
gunicorn
